// HealthCareManagement-main/src/app/models/appointment.ts
export class Appointment {
  patientname: string = '';
  patientid: string = '';
  email: string = '';
  doctorname: string = '';
  specialization: string = '';
  date: string = '';
  age: string = '';
  gender: string = ''
  problem: string = '';
  slot: string = '';
  appointmentstatus: string = 'false';
  admissionstatus: string = 'false';

  constructor() { }
}